requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.6
six==1.11.0
pyyaml>=3.13
pyserial
pyredis
